+++ zerver/lib/event_types.py
@@ -627,6 +627,7 @@ class GroupSettingUpdateData(GroupSettingUpdateDataCore):
     can_summarize_topics_group: int | UserGroupMembersDict | None = None
     direct_message_initiator_group: int | UserGroupMembersDict | None = None
     direct_message_permission_group: int | UserGroupMembersDict | None = None
+    workplace_users_group: int | UserGroupMembersDict | None = None
 
 
 class PlanTypeData(BaseModel):